pyproject.toml
[tool.poetry]
name = "coursework-django-rest-framework"
version = "0.1.0"
description = "Django REST Framework Habit Tracker"
authors = ["Vasily Umelyantsev <tanec_991@mail.ru>"]
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.7"
djangorestframework = "^3.16.1"
djangorestframework-simplejwt = "^5.5.1"
django-cors-headers = "^4.9.0"
drf-yasg = "^1.21.11"
django-celery-beat = "^2.8.1"
celery = "^5.5.3"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.0.0"
requests = "^2.32.5"
gunicorn = "^21.2.0"
redis = "^5.0.0"

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
isort = "^7.0.0"
mypy = "^1.18.2"
flake8 = "^7.3.0"
coverage = "^7.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
